# adapters/http/openapi.yaml
openapi: 3.1.0
info:
  title: dagworker HTTP/JSON worker protocol
  version: "1.0.0"
  description: >
    Hand-written contract (docs/research/14-http-json-worker-protocol.md;
    ADR-0037) for workers that participate in a dagworker DAG over plain
    HTTP/JSON instead of embedding the Go library or a gRPC stub. Resources
    (scopes, nodes, edges) follow standard CRUD verbs. The two operations
    where the server, not the client, decides the outcome — claim and every
    lease acknowledgement — are AIP-136 colon-suffixed custom methods.
...
servers:
  - url: /v1
paths:
...
  /scopes/{scope}:cancel:
    parameters:
      - $ref: "#/components/parameters/scope"
    post:
      summary: Cancel every non-terminal node in the scope
      operationId: cancelScope
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scope" }
        default: { $ref: "#/components/responses/Problem" }
  /scopes/{scope}/nodes:
    parameters:
      - $ref: "#/components/parameters/scope"
    get:
      summary: List nodes, keyset-paginated
      operationId: listNodes
      parameters:
        - name: status
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/Status" } }
          style: form
          explode: false
        - name: kind
          in: query
          schema: { type: array, items: { type: string } }
          style: form
          explode: false
        - name: cursor
          in: query
          description: Opaque continuation token from a previous page's next_page_token. There is no offset parameter — keyset pagination only.
          schema: { type: string }
        - name: page_size
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items: { $ref: "#/components/schemas/Node" }
                  next_page_token: { type: string }
        default: { $ref: "#/components/responses/Problem" }
  /scopes/{scope}/nodes/{node}:
    parameters:
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/node"
    put:
      summary: Create a node (client-chosen ID). Idempotent — a byte-identical re-PUT is a no-op.
      operationId: putNode
      parameters:
        - name: If-Match
          in: header
          description: Optional. "*" or a prior ETag; rejects the write with 412 if the node's current representation does not match.
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateNodeRequest" }
      responses:
        "200": { description: "Updated (identical spec re-PUT)", content: { application/json: { schema: { $ref: "#/components/schemas/Node" } } } }
        "201": { description: "Created", content: { application/json: { schema: { $ref: "#/components/schemas/Node" } } } }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        default: { $ref: "#/components/responses/Problem" }
    get:
      summary: Fetch one node
      operationId: getNode
      parameters:
        - name: If-None-Match
          in: header
          schema: { type: string }
      responses:
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "304": { description: "Not Modified" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      summary: Remove a node (only if not claimed; see ?cascade for successors)
      operationId: deleteNode
      parameters:
        - name: cascade
          in: query
          schema: { type: string, enum: [reject, detach, fail], default: reject }
        - name: If-Match
          in: header
          schema: { type: string }
      responses:
        "204": { description: "No Content" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    scope:
      name: scope
      in: path
      required: true
      schema: { type: string, maxLength: 255 }
    node:
      name: node
      in: path
      required: true
      schema: { type: string, maxLength: 255 }
...
  responses:
    Problem:
      description: An error, as an RFC 9457 Problem Details object.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PreconditionFailed:
      description: If-Match did not match the resource's current ETag.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Status:
      type: string
      enum: [new, in_progress, success, error]
    Reason:
      type: string
      enum: [none, worker_error, timeout, upstream_failed, skipped, cancelled, removed]
    Trigger:
      type: string
      enum: [all_success, all_done, none_failed, none_failed_min_one_success, always]
    Node:
      type: object
      properties:
        name: { type: string, description: 'AIP-122 resource name, "scopes/{scope}/nodes/{id}"' }
        scope: { type: string }
        id: { type: string }
        kind: { type: string }
        status: { $ref: "#/components/schemas/Status" }
        reason: { $ref: "#/components/schemas/Reason" }
        message: { type: string }
        attempt: { type: integer }
        priority: { type: integer }
        trigger: { $ref: "#/components/schemas/Trigger" }
        retry_max_attempts: { type: integer }
        retry_base_delay: { type: string, description: "Go duration string, e.g. \"1s\"" }
        retry_max_delay: { type: string }
        payload_encoding: { type: string, enum: [base64] }
        payload: { type: string, contentEncoding: base64 }
        labels: { type: object, additionalProperties: { type: string } }
        seq: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateNodeRequest:
      type: object
      properties:
        kind: { type: string }
        payload_encoding: { type: string, enum: [base64] }
        payload: { type: string, contentEncoding: base64 }
        priority: { type: integer }
        trigger: { $ref: "#/components/schemas/Trigger" }
        max_attempts: { type: integer }
        retry_base_delay: { type: string }
        retry_max_delay: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        dependencies: { type: array, items: { type: string } }
    ScopeConfig:
      type: object
      description: All fields optional; zero/absent means "use the conservative fallback" (docs/spec/01-contract.md §11).
      properties:
        default_lease_timeout: { type: string }
        min_lease_timeout: { type: string }
        max_lease_timeout: { type: string }
        max_attempts: { type: integer }
        retry_base_delay: { type: string }
        retry_max_delay: { type: string }
        terminal_retention: { type: string, description: "0 (absent) means never collect" }
        max_subscriber_lag: { type: string, description: "0 (absent) means unbounded" }
        max_in_flight: { type: integer }
        payload_cap: { type: integer }
        max_batch_size: { type: integer }
        sweep_batch_size: { type: integer }
        sweep_interval: { type: string }
        partition_count: { type: integer }
    ScopeStats:
      type: object
      properties:
        total: { type: integer }
        blocked: { type: integer }
        scheduled: { type: integer }
        ready: { type: integer }
        in_progress: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        sealed: { type: boolean }
        complete: { type: boolean }
        cursor: { type: integer }
    Scope:
      type: object
      properties:
        name: { type: string }
        config: { $ref: "#/components/schemas/ScopeConfig" }
        stats: { $ref: "#/components/schemas/ScopeStats" }
...
    Problem:
      type: object
      description: RFC 9457 application/problem+json. Extension members (e.g. current_etag, field, size/cap) vary by "type".
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [type, title, status]